requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
urllib3==1.26.13
pytest==7.2.0
moviepy==1.0.3